ai-proma/requirements.txt
fastapi==0.116.1
anthropic==0.67.0
numpy==2.3.3
openai==1.107.1
psycopg==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
regex==2025.9.1
requests==2.32.5
SQLAlchemy==2.0.43
tiktoken==0.11.0
tqdm==4.67.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
databases==0.9.0
asyncpg==0.30.0
PyJWT==2.10.1
passlib==1.7.4
email-validator==2.3.0
litellm==1.77.0
google-adk==1.14.0
google-generativeai==0.8.5
google-cloud-storage==2.19.0
google-genai==1.36.0
vertexai
